move isKeyOwnerOrLockVerifier to metadataOperation
