+++ locksmith/src/operations/metadataOperations.ts
@@ -3,10 +3,21 @@ import { LockMetadata } from '../models/lockMetadata'
 import Metadata from '../../config/metadata'
 import KeyData from '../utils/keyData'
 import { getMetadata } from './userMetadataOperations'
+import { Web3Service } from '@unlock-protocol/unlock-js'
+import networks from '@unlock-protocol/networks'
+import { Verifier } from '../models/verifier'
+import Normalizer from '../utils/normalizer'
 
 const Asset = require('../utils/assets')
 
 const baseURIFragement = 'https://assets.unlock-protocol.com'
+
+interface IsKeyOrLockOwnerOptions {
+  userAddress?: string
+  lockAddress: string
+  keyId: string
+  network: number
+}
 
 export const updateKeyMetadata = async (data: any) => {
   try {
@@ -137,4 +148,33 @@ const defaultMappings = (address: string, host: string, keyId: string) => {
   // Append description
   defaultResponse.description = `${defaultResponse.description} Unlock is a protocol for memberships. https://unlock-protocol.com/`
   return defaultResponse
+}
+
+export const isKeyOwnerOrLockVerifier = async ({
+  userAddress,
+  lockAddress,
+  keyId,
+  network,
+}: IsKeyOrLockOwnerOptions) => {
+  if (!userAddress) {
+    return false
+  }
+  const web3Service = new Web3Service(networks)
+  const loggedUserAddress = Normalizer.ethereumAddress(userAddress)
+
+  const isVerifier = await Verifier.findOne({
+    where: {
+      lockAddress,
+      userAddress,
+      network,
+    },
+  })
+
+  const keyOwner = await web3Service.ownerOf(lockAddress, keyId, network)
+
+  const keyOwnerAddress = Normalizer.ethereumAddress(keyOwner)
+
+  const isKeyOwner = keyOwnerAddress === loggedUserAddress
+
+  return isVerifier?.id || isKeyOwner
 }
